Add first attempt at reachability_graph
Add an initial version of `reachability_graph` that is able to process
a `boost::adjacency_list` and calculate the number of unique paths
between two vertices.  It has an implementation of the `reachable_from`
method that was also required from the algorithm sketch in `README.md`.
It does not yet implement `sources_that_can_reach` though.

`boost` has been added as a Conan dependency and the minimum C++ version
changed to C++20 in order to avoid writing `typename` to disambiguate
types and non-types in some contexts.