Add proof of concept for unit testing
Create a static library `common` to move out our core functionality.
This way it can be linked in to a new unit test executable.

Add a `tests` target that depends on `common` and also `gtest` (via
conan).  Add a unit test that sets up an in-memory file system so we can
test our frontend action tools.