Fix printing empty paths
If we try to get a relative path from `./` from absolute paths then the
result is an empty string.  This was found when running
`includeguardian` across our own codebase.

The fix is to manually remove "./" for unit tests in quite an ugly way.
However, clang probably shouldn't be giving us forward slashes on
Windows anyway as a directory separator.  Most likely these sort of
issues will arise again and we may have to fix it upstream in clang.