Fix a crash with an empty pragma string
When running `includeguardian` across our own codebase I encountered a
crash where `getSpelling` will return an empty string with a `nullptr`.
This causes a crash in `std::equal`.  It is much simpler to use
`startsWith`.

I am unsure why we get an empty pragma string at the moment and will
have to investigate later on.