Add get_total_cost component
Add a `get_total_cost` component that will find the total file size of
each source after preprocessing.  This may not be entirely accurate as
this assumes that every file included has a unique include guard.

In the future it may be better to get this information in the
`build_graph` step as we may be able to get the size of the exact file
that clang is left with after preprocessing.