Add parallelism to find_expensive_includes
Add some reasonably naive parallelism to `find_expensive_includes`.
When running on itself, it goes from ~3000ms to find all expensive
includes to ~800ms on my laptop - a reasonbly good saving!

After this the main bottleneck is to run the preprocessor over all of
the files (2300ms).  A future change should look at parallelising this
step.