Minor refactor to DFS search state
Change from `std::vector<T>` to `std::unique_ptr<T[]>` since we are a
fixed size and we can avoid doing the initial `fill` in the constructor.
A trivial performance improvement but more important now that we
construct these helper objects a lot more frequently since the change to
make the code parallel.

Replace the RAII on clearing the stack with an assertion that it is
clear and add an explicit `m_stack.clear()` in the one extra place we
need it.