Add -ferror-limit=0
When running includeguardian across chromium there are a lot of errors
for missing defines that presumably are generated dynamically.  Most of
these defines probably do not contribute significantly to an accurate
list of expensive includes.  So for the sake of easy-of-testing I am
hardcoding the limit to be infinite at the moment.

Though this is possible to pass in as `-extra-arg="-ferror-limit=0"`, I
don't want to have to do this any time I test a large codebase.