Allow missing includes
To support a "best effort" attempt at estimating the most expensive
includes, allow include directive to missing files to be a warning
rather than an error.  This will allow us to run against a new project
and get a list of all missing files.  We can fixup the missing ones that
are most likely to be large (e.g. boost) and ignore smaller, less
frequently used libraries.

We don't need to inject `-ferror-limits=0` as a command line argument
now that we are handling the diagnostics directly and can call
`setErrorLimit`.