Fix parallel edges
When building the graph we would include parallel edges between the same
nodes if they were reachable by multiple source files.

This was fixed by having a lookup between a file's unique ID that
persisted across preprocessing different files.  We also added a stack
of the current files open in order to avoid more lookups.

There are a few quirks in clang's preprocessor that I had to work
around.  I am reasonably confident that there are good reasons that
clang works like this, but right now I don't know what they are and I
don't want to be blocked.  These issues are documented in the code.

A unit tests have been added with multiple sources to test.