List out expensive files
Add more functionality to includeguardian to list out files that are
particularly large and included my multiple sources.  This can indicate
that splitting up these files into smaller ones would have a impact in
build times.

This commit fixes a few issues where we ran out of stack space.  One in
an `operator<<` which had infinite recursion, and one in the
implementation in `reachability_graph` where we could recurse enough to
blow the stack.