Change from compilation database to directory for input
It is a bit overkill to use a compilation database for input because we
don't actually need to compile anything.  From my personal experience
testing this out, it's much nicer to be able to pass in a directory path
and instead take all source files from within there.

We can pass in additional include directories with the `--i` flag and
pass in the directory with `--dir`.