Fix #pragma override_file_size
`#pragma override_file_size(N)` was not working as the `StringRef`
returned from `getSpelling` would return only the `#` character.
Instead we make the assumption (which I am pretty sure is true) that all
source file buffers are null-terminated for performance reasons, e.g.
you can check for `'\0'` instead of being at the end of the buffer.

Update the unit test to actually compare the `file_node` objects to
catch this issue.