Change cost from file size to # of preprocessor tokens
I have no proof, but I have read a few places online that the cost to
syntax check a file is correlated more to the number of preprocessing
tokens compared to the raw file size.  I have no proof of this, but it
makes sense that something like comments or whitespace take almost no
time compared to actual C++ code.

In the future we need benchmarks to prove this correlation (and possibly
find a better metric).