Add analysis for expensive headers
Add analysis that gives the benefit of removing all references to a
particular file.  This is not quite finished as we never update the
`sources_count` variable.  We also would like to tweak the analysis a
little bit so that we either only remove includes to the file from
headers or move includers from headers into their respective sources.
I think this is a more accurate view of the world, that you can't
usually remove the use of a header completely, but instead push its use
into a source file.

This code also gets quite ugly when testing as the cost is split into
preprocessing tokens and file size.  A future change will create a
`cost` component and combine them.