Link header and source files
In preparation for a future analysis we need to know if a header file
has an associated source, and vice versa.

For example, for a source file that contains very little, it may be
worth while to move that code into the header in order to reduce a whole
source compilation - perhaps if the header is quite large.

Also in reverse, we should improve `find_expensive_headers` so that it
doesn't just remove all includes of a particular header file, instead it
should recommend moving it to the source file if there is one.  And if
there isn't, we should understand the cost of creating it.

Both of these optimizations (and I'm sure others) require knowledge of
the header-source connection.  Right now we guess that it's going to be
the first include that shares the filename stem.