Add count of incoming includes
Inside `file_node`, keep track of the number of times this file is
included.  Compared to changing `Graph` to store the reverse edges using
`boost::bidirectionalS`, it is a relatively cheap metric to keep the
number of incoming includes.  This allows us to skip all sources inside
`find_expensive_headers`.