Update find_expensive_headers
Change the algorithm to find the cost of moving all inclusions for a
file to their corresponding source files (or create one if necessary).
This is a more accurate representation of how something like this could
work rather than outright removal of certain inclusions.

I have had to change the `token_count` type in `cost` from `unsigned` to
`int` to make subtraction less prone to underflow.