Add list_included_files
Add a way to list out all of the files along with how many source files
directly or indirectly include these files.

This was added so we can get a rough idea of the distribution of include
cost among files in preparation for writing some analysis to recommend
an idea precompiled header.  In order to do this we probably want an
rough guess of what percentage reduction we should be aiming for (since
we can get a 100% reduction with precompiled headers, but it may
generate a file that is too large).