Add suggestions for what to add to a precompiled header
Reuse `find_expensive_headers` to list files that would be most
beneficial in adding to a precompiled header.  In this case we don't
need to consider the cost to move the includes to the respective source
file because we assume for now that there is no cost to including a file
that is part of a precompiled header.

Update our precompiled header based on the recommendations we get

```
Precompiled header suggestions
8,223,955 (9.13%) adding "boost\\graph\\adjacency_list.hpp" to a precompiled header
1,604,442 (2.83%) adding "chrono" to a precompiled header
1,274,544 (1.41%) adding "charconv" to a precompiled header
1,173,856 (1.30%) adding "algorithm" to a precompiled header
480,176 (0.53%) adding "boost\\config.hpp" to a precompiled header
470,516 (0.52%) adding "memory" to a precompiled header
268,489 (0.30%) adding "functional" to a precompiled header
175,080 (0.19%) adding "iterator" to a precompiled header
```

I am surprised that there aren't more standard library headers such as
`<vector>`, `<filesystem>`, or `<type_traits>`.  This will be something
to investigate later on.