Add support for precompiled headers
Add basic support for a precompiled header with the `.pch` extension.
This will cause all files with that extension to effectively have zero
cost when performing are analysis.

Future changes should allow injecting an include for every source file
so we can support precompiled headers in our own project (and most
likely everyone elses as this is the best way to do it).