Refactor build_graph
In preparation for adding forced include support, we refactor how
`build_graph` is implemented by taking out the core lexing into a
`process` method.

This commit also adds in a rough idea of how we are going to process the
forced includes.  The remaining change is to pass in a list of forced
includes into `process` and make sure they are always connected.  After
that we should add unit tests.