Add support for forced includes
Allow the user to pass in zero or more `--forced-include` paths to C/C++
files that will be included at the beginning of each source file.  This
will most likely be used to inject precompiled headers.

I have tested this out on the `includeguardian` project and found that
I could include `<filesystem>` in my precompiled header to decrease the
amount of code seen by 5% for a full rebuild.