Add recommend_precompiled
Add a component that is dedicated to recommended what files should
appear in the precompiled header.  This was previously done by
`find_expensive_headers`, but for several reasons that doesn't give the
correct answer.

There are no unit tests added for this component.  I would like to take
this opportunity to reorganize the existing tests so it is easier to
share the input graphs among each test.

This was tested against our own source code and our list of precompiled
headers altered to match the suggestion.