Use clang::Preprocessor
Conditional compilation needs to be added because we end up with a lot
of warnings about missing Linux files on Windows and vice versa.  This
is quite a difficult problem to solve ourselves with `clang::Lexer` and
it would be a better idea to move back to using `clang::Preprocessor`
and the `PPCallbacks` interface.

I moved away from `PPCallbacks` before in
0c13da4c364a68a72b8098d5cc0f03967d9162fb for a few reasons.  Summarizing
here:

  1. It was difficult to get the relative path of an include with
     relation to where it was found in the list of include directories.
  2. Performance
  3. It was hacky to work around some of the quirks of `PPCallbacks` and
     the order in which methods would be invoked.

I believe that all of these are fixable and are better for this project
than trying to implement a preprocessor myself.

  1. This can be replicated more accurately using `clang::HeaderSearch`,
     but I think the `rel_to_one_of` function does a good job of finding
     a reasonable file path.  In the future we may want to stick with
     absolute paths because it makes it easier for the user to jump
     straight to that file.
  2. By using the `VirtualFileSystem` we can replace the contents of
     files that we have already preprocessed with a simplified version
     that only includes macro definitions and perhaps other includes (to
     implement `__has_include`).  This should only be a bit slower than
     what we do today but it also is closer to being feature-complete.
  3. These were previously worked around so aren't actually an issue.  I
     understand a few things better like the `nullptr` `clang::File *`
     being from the predefines.  Perhaps there are otherways to remove
     the other hacks.

It is worth expanding on the idea for 2.  Here we make an assumption
that if a header with an include guard is included, then any macro
definitions that change the output of the preprocessor are always
defined in the exact same way each time this header is included.  I
cannot think of a valid reason why this wouldn't be the case.

For headers that do not have an include guard, then we shouldn't replace
its contents the next time it is included.  These files most likely
contain the `X`-macro idiom and instead we can merge its cost into the
includee.

Changing the subject, another option is to not use clang for the
preprocessor but to use something else like boost's wave library.  I
think the wave library is more modular and documented better for end
users, but it does not have a virtual file system support so
performance would be an issue.  It also will (most likely) never be as
cutting edge as clang will be.  All new compiler features that are
likely to affect us (e.g. modules) will be supported very quickly in
clang.

This is passing all unit tests, plus an additional test I wrote using
macros that checks to see that the cost of a file is the number of
preprocessing tokens **after** expanding macros.

However, this is crashing when running against the IncludeGuardian code
base.  This will be fixed in a later change as this commit is already
large.