Fix use of PPCallbacks
My understanding of `PPCallbacks` was very flawed.

`FileID` passed in `FileChanged` is always the file that we are going
to, even if `Reason == ExitFile`.  In the case of `ExitFile` we need to
use the `OptionalPrevFID` value instead.  This fixed the remaining 2/3
issues I had with `PPCallbacks`:

  1. `ExitFile` was called too frequently for the main source file - it
     actually was being called for exiting the includes of the main
     source file going **back** to the source.
  2. We didn't get any `ExitFile` for files that didn't have include
     directives - this was because `FileID` would always be the ID of
     the file we jumped back to instead of from.

The final issue with `PPCallbacks` was due to `nullptr` values coming
back, but I have already understanding that these are due from the fake
file containing the predefines.