Improve how token count is recorded
We want to move back to clang's tooling over manually creating a
`clang::Preprocessor` object.  The reason being is that this interface
is designed for external use and is set up from the get-go.  In
particular I am getting assertions when running IncludeGuardian on my
own code and I assume it's because my set up is not 100% correct.

In order to do this we have to avoid doing anything within the loop
calling `Lex`.  Right now this increments the correct token count.  This
changes the code to record the token count when we move between files.