Remove the need to examine the list of include directories
As part of the continuing work to move back to using clang Tooling, we
want to remove the need for `IncludeScanner` to have access to anything
other than the `clang::Preprocessor` object.

This commit removes the list of include directories passed in, which may
differ between source files if later we allow the use of a compilation
database.  Previously the code would use the include directories to
generate a reasonable relative filename for what we include.  For
example if we had `boost/example.hpp` include `"foo.hpp"` we would want
the relative path to be `boost/foo.hpp` so the user knows that it came
from the `boost` folder that they most likely added as an include
directory.

To allow this, we keep concatenating the relative paths of the includes
that we see, resetting this to the parent path of any `<>` include that
we come across.