Handle unguarded include files
Some include files do not guard themselves with `#pragma once` or
include guards either because they are designed with this in mind (e.g.
X macros) or accidental.  In this case we attribute the cost of
processing this file from the includer since with X macros, the cost
will differ dependening on what you define `X` to be.