Move back to ClangTool
We finally move back to `ClangTool` and `PPCallbacks` after swapping to
manually lexing the source files.  This should prove to be more robust
to API changes in `Clang`.

Future changes will need to swap out `fully_processed` files in our
`VirtualFileSystem` with only `#define` lines in order to speed up
building the graph when we have a large number of sources.