Fix an issue with unexpected nullptr
When extracting the text from a `#pragma`, we can get an empty
`clang::StringRef`, which means `data()` can return `nullptr`, causing a
seg fault in our call to `std::equal`.

Swap this to `std::mismatch` to avoid this error.  Note that we assume
that if the string isn't `nullptr`, it is at least `null`-terminated
later on - a fact I think still holds.

I wasn't able to replicate in unit test with a `"#pragma\n"` line in the
unit test.  Usually I would investigate a bit harder, but I think time
is better spent getting a MVP rather than 100% test coverage.

With this, we are back once again to being able to run IncludeGuardian
against our own code base :)