Support -I and -isystem
Allow both ways to add include directories in the command line.  `-I`
will mean all files founds in this directory are counted as internal and
we will perform analysis on them.  `-isystem` will mean all files found
in this directory are counted as external and we'll skip analysis on
them.