Add benchmark scaffolding
Add a `benchmarks.m.cpp` file and project with a Google benchmark
dependency.

This is preparation to start improving performance of the
preprocessing and analyses:

  1. Once we have processed a file that is correctly guarded, we can
     swap it out with a smaller file that only defines the same macros
     and has `#pragma`s to fake the file size and token count.  This
     should get us back closer to the speed that we had when we were
     lexing manually.
  2. Detect earlier when a particular file/include/line is never going
     to hit our percentage threshold and exit early in our analysis.
  3. For a particular file(s) we are analysing, walk back up the tree to
     find only those source files that depend on it and then DFS only
     from those.

When looking at the `CMakeLists.txt` file, I also discovered that the
main `includeguardian` project was not using the precompiled header from
`common`, which would have slowed down compilation.