Track the number of includers that are external
One we have the number of external files that include us, we can sum
this with the internal number to get the total number of files that
include us.  This will allow us to skip some false positives when
complaining about files without guards, if those files are only ever
included once.