Use bidirectionalS with boost graph
Make our map allow iteration of the in edges in our graph.  In this
commit it is used to replace summing the internal and external counts of
incoming edges, in the future it will allow us to backtrack from a file
up to all of the sources that includes it so we can do less work during
our analysis.