Support compilation databases
Support the `-p` option to specify a compilation database in which to
grab all of the compilation flags (and also the other flags that clang
tools usually support).

We do this by using the `from_compilation_db` method when building a
graph instead of `from_dir`.  If we are specified `--dir` then we'll
create a fake compilation database and pass it in.