Partially fix `smaller-file-opt`
The previous issue was that `clang::FileManager *` is reused across
translation units and this has a cache of paths against `clang::File *`.
It meant if we swapped out files in our `VirtualFileSystem`, we would
not even try to look them up because we would do the search in the cache
first.

This change creates a new `clang::FileManager` for each translation unit
and only swaps the file contents at the end of processing each source.

There are still some discrepancies with the output between runs with the
optimization on and off.  Though they are minor, they still need to be
looked at with some importance.  In particular there are quite egregious
issues such as the total file size being incorrect.