Allow (de)serialization of the graph
In order to avoid building the graph repeatedly and to help diagnose the
issue where the `smaller-file-opt` flag doesn't quite generate the
correct result, we have the `--save` and `--load` flags that can
(de)serialize our graph to disk.

In the future we could allow our analysis recommendations to be able to
modify the graph and save it back.  For example, if we recommend
removing one include directive and the user manages to do it, instead of
rebuilding the entire graph from the source, we can just change the
graph directly and perhaps save it to disk again.