Large fix of several issues
When we run IncludeGuardian across our own code base with and without
the small file optimization, we get different serialized outputs of the
graph.

This was due to several different issues, specifically around
non-guarded files changing which their includes whenever they were
included, and the fact that we did not rollup correctly the cost of
unguarded files into the nearest guarded ancestor, if there were
unguarded nodes in the middle.

This change adds in a few more tests, fixes others, and afterwards
we get the same graph when running against the IncludeGuardian code base
with and without the optimization.