Fixed UB when including `<windows.h>`
When including `<windows.h>` we get a path of `C:\path\to\windows.h` but
`<windows.h>` eventually ends up including a file that includes
`"windows.h"`, however clang gives us `C:\path\to/winows.h` because in
`HeaderSearch.cpp` it concatenates paths for non-angled includes with a
forwardslash regardless of system.

This means that we end up replacing the same file twice when
`--smaller-file-opt=true` and it eventually causes a crash.

This has been fixed by an `OverwriteFileSystem` that replaced files by
`UniqueID` and delegates looking up a path to the underlying file
system.  So no matter how many different paths map to the same file on
any system, we should be replacing this only once.

To test this the following setup failed.

   test.cpp
   #include <windows.h>

   test2.cpp
   #include <windows.h>