Fix `find_expensive_headers`
When running this against the public BDE repository we do not see
a recommendation to make `<intrin.h>` an implementation detail, even
though this would result in a huge 17% saving.

The algorithm was incorrect in a few ways and was rewritten to actually
build the "after" graph and check the cost of this.  It can be written
in a more performant way I'm sure, but that can be done later on.

More unit tests were written.