Output console text earlier
When running expensive analysis, we only output something once it has
finished.  By outputting the comment and the name before starting the
analysis, we can give faster feedback to the user and they can see what
they are waiting on.

Include the post-analysis sorting and filtering in the calculation of
how much time it took.