Fix `find_expensive_headers`
When investigating whether a header can be made private, we did not
exclude modifications to external files.

For example, the header `<minintrin.h>` was showing up when analysing
public BDE code and said that if these 2 directives were removed we
would save 16%.  However, when I removed these includes it saved almost
nothing.  The problem was that the analysis was assuming that we could
edit `<intrin.h>` to remove the include of `<minintrin.h>`.