Fix serialization and deserialization
Our implementation of some of the `boost::serialization` stuff wasn't
done correctly.  I don't really understand what went wrong but it's
simpler now and it works.

I also completely missed that I must have removed the bit of code in
`includeguardian.cpp` that read in if we passed in `--load`!

I've added some unit tests for serialization/deserialization that pass
now.

The remaining issue is that we do not serialize/deserialize the set of
sources so when we pass in `--load` it does nothing.