Fix --load not doing anything
When passing `--load` we didn't do anything because we don't take our
sources from the underlying file.  We could allow the standard `**/*`
argument to look in directories for sources, but we don't know whether
we have the source files that match.  It is much easier to serialize the
entire `build_graph::result` object and deserialize it on `--load`.