Workaround assertion occurring in unit tests
There is an interesting assert that occurs only during unit tests
when we have a file `idempotent.hpp` that would be identical
to the `enable_replace_file_optimization` replacement.  This is
because in both unit tests and when doing the optimization we
use `InMemoryFileSystem`.  This generates the `UniqueID` for the
file based on its file and content, so in our case we would get
identical `UniqueID` for the original and replacement, even though
they occur in different `InMemoryFileSystem` instances.  This hits
an assert in our code `assert(new_id != id)` in `build_graph.cpp`
where we want to check that the replacement has occurred.

This is "fixed" by putting a comment at the top of each replacement
file that shouldn't occur during unit tests.

This was found on Linux-only, perhaps because of line endings being
different in Windows.