Fix most mobile website issues
The mobile website was very difficult to use and it cut off a lot of
text.  This allieviates most (all?) of the problems.  I will have to
investigate a more robust way of preventing the content of the
navigation bar or article content from overflowing its parent container.
When this happens the website is very hard to read.